BMI report generator
the project is based on a user input player bmi generator which output is a pdf script. there is two admin panel one is for create player using player name and position and other is creating a team for tournament. both user inface can save bmi by selecting player name and give input player height as a feet and weight as a kg.